+++ packages/client-node/__tests__/unit/node_result_set.test.ts
@@ -13,7 +13,7 @@ describe('[Node.js] ResultSet', () => {
   })
 
   it('should consume the response as text only once', async () => {
-    const rs = makeResultSet()
+    const rs = makeResultSet(getDataStream())
 
     expect(await rs.text()).toEqual(expectedText)
     await expectAsync(rs.text()).toBeRejectedWith(err)
@@ -21,17 +21,15 @@ describe('[Node.js] ResultSet', () => {
   })
 
   it('should consume the response as JSON only once', async () => {
-    const rs = makeResultSet()
-
+    const rs = makeResultSet(getDataStream())
     expect(await rs.json()).toEqual(expectedJson)
     await expectAsync(rs.json()).toBeRejectedWith(err)
     await expectAsync(rs.text()).toBeRejectedWith(err)
   })
 
   it('should consume the response as a stream of Row instances', async () => {
-    const rs = makeResultSet()
+    const rs = makeResultSet(getDataStream())
     const stream = rs.stream()
-
     expect(stream.readableEnded).toBeFalsy()
 
     const result: unknown[] = []
@@ -50,10 +48,8 @@ describe('[Node.js] ResultSet', () => {
   })
 
   it('should be able to call Row.text and Row.json multiple times', async () => {
-    const rs = new ResultSet(
+    const rs = makeResultSet(
       Stream.Readable.from([Buffer.from('{"foo":"bar"}\n')]),
-      'JSONEachRow',
-      guid(),
     )
     const allRows: Row[] = []
     for await (const rows of rs.stream()) {
@@ -67,14 +63,21 @@ describe('[Node.js] ResultSet', () => {
     expect(row.json()).toEqual({ foo: 'bar' })
   })
 
-  function makeResultSet() {
-    return new ResultSet(
-      Readable.from([
-        Buffer.from('{"foo":"bar"}\n'),
-        Buffer.from('{"qaz":"qux"}\n'),
-      ]),
-      'JSONEachRow',
-      guid(),
-    )
+  function makeResultSet(stream: Stream.Readable) {
+    return ResultSet.instance({
+      stream,
+      format: 'JSONEachRow',
+      query_id: guid(),
+      log_error: (err) => {
+        console.error(err)
+      },
+    })
+  }
+
+  function getDataStream() {
+    return Readable.from([
+      Buffer.from('{"foo":"bar"}\n'),
+      Buffer.from('{"qaz":"qux"}\n'),
+    ])
   }
 })