+++ packages/client-common/src/connection.ts
@@ -29,6 +29,7 @@ export interface ConnBaseQueryParams {
   abort_signal?: AbortSignal
   session_id?: string
   query_id?: string
+  auth?: { username: string; password: string }
 }
 
 export interface ConnInsertParams<Stream> extends ConnBaseQueryParams {